[ANDROSDK-1964] Move AttributeFields to network package
